OSX bug fix, tweaks
Fixed OSX button display bug, tweaked a function to avoid unnecessary
declaration of a variable and removed the global ”LEFT_SIDEBAR_WIDTH”
as it is no longer needed. My IDE also fixed the indentation in the
files that I worked on. Would be good if someone could test this bug
fix cross platform.